Add experimental version of plugin system test.

+++ src/plugins/mvcc_delete_plugin.cpp
@@ -78,23 +78,26 @@ void MvccDeletePlugin::_logical_delete_loop() {
 void MvccDeletePlugin::_physical_delete_loop() {
   std::unique_lock<std::mutex> lock(_mutex_physical_delete_queue);
 
-  TableAndChunkID table_and_chunk_id = _physical_delete_queue.front();
-  const auto& table = table_and_chunk_id.first;
-  const auto& chunk = table->get_chunk(table_and_chunk_id.second);
+  if(_physical_delete_queue.size()) {
+    TableAndChunkID table_and_chunk_id = _physical_delete_queue.front();
+    const auto& table = table_and_chunk_id.first;
+    const auto& chunk = table->get_chunk(table_and_chunk_id.second);
 
-  DebugAssert(chunk != nullptr, "Chunk does not exist. Physical Delete can not be applied.");
+    DebugAssert(chunk != nullptr, "Chunk does not exist. Physical Delete can not be applied.");
 
-  if (chunk->get_cleanup_commit_id().has_value()) {
-    // Check whether there are still active transactions that might use the chunk
-    bool conflicting_transactions = false;
-    auto lowest_snapshot_commit_id = TransactionManager::get().get_lowest_active_snapshot_commit_id();
-    if (lowest_snapshot_commit_id.has_value()) {
-      conflicting_transactions = chunk->get_cleanup_commit_id().value() < lowest_snapshot_commit_id.value();
-    }
+    if (chunk->get_cleanup_commit_id().has_value()) {
+      // Check whether there are still active transactions that might use the chunk
+      bool conflicting_transactions = false;
+      auto lowest_snapshot_commit_id = TransactionManager::get().get_lowest_active_snapshot_commit_id();
+      
+      if (lowest_snapshot_commit_id.has_value()) {
+        conflicting_transactions = chunk->get_cleanup_commit_id().value() < lowest_snapshot_commit_id.value();
+      }
 
-    if (!conflicting_transactions) {
-      _delete_chunk_physically(table, table_and_chunk_id.second);
-      _physical_delete_queue.pop();
+      if (!conflicting_transactions) {
+        _delete_chunk_physically(table, table_and_chunk_id.second);
+        _physical_delete_queue.pop();
+      }
     }
   }
 }
